Separated program initializer concerns
