+++ TiempoDev.DataTypeSamples.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27703.2047
+# Visual Studio 14
+VisualStudioVersion = 14.0.25420.1
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TiempoDev.DataTypeSamples", "TiempoDev.DataTypeSamples\TiempoDev.DataTypeSamples.csproj", "{036E85EB-F6A8-4A37-974D-FDFEC2C0A481}"
 EndProject
