+++ TiempoDev.DataTypeSamples/Program.cs
@@ -1,41 +1,10 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading;
-using System.Threading.Tasks;
 using TiempoDev.DataTypeSamples.Interactive;
-using TiempoDev.DataTypeSamples.Utils;
 
 namespace TiempoDev.DataTypeSamples
 {
-    public class Program
+    partial class Program
     {
-        internal static readonly TerminalFormatter _formatter;
-
-        static Program()
-        {
-            try
-            {
-                TerminalFormatter.EnableWindowsVirtualTerminalSequences();
-            }
-            catch (Exception)
-            {
-                // Ignore
-            }
-
-            try
-            {
-                TerminalFormatter.EnableWindowsVirtualTerminalSequences(stdErr: true);
-            }
-            catch (Exception)
-            {
-                // Ignore
-            }
-
-            _formatter = new TerminalFormatter();
-        }
-
         static void Main(string[] args)
         {
             try
