+++ TiempoDev.DataTypeSamples/TiempoDev.DataTypeSamples.csproj
@@ -54,6 +54,7 @@
     <Compile Include="Operators.cs" />
     <Compile Include="PrimitiveTypes.cs" />
     <Compile Include="Program.cs" />
+    <Compile Include="Program.Initializer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Interactive\SampleRunner.cs" />
   </ItemGroup>
